requirements.txt
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
chainer==1.19.0
Pillow==4.0.0
beautifulsoup4==4.5.3
chardet==3.0.2
